Centralizing 'banner' and adding valgrind to test coverage
Adding a zero-length config file example to trigger the EINVAL
error on the mmap call to test that error path.

Adjusted 'config_parsed' variable to be visible to allow resetting
it to test multiple config files.

All error messages pass through syslog now, with a central file of
'banners' to allow for future common strings to be added as needed
for the program.

tests/parser/parser.sh automagically adds valgrind testing if that
tool is installed.